refactor(clusterview): split into color/list/detail/compare files
clusterview.go was 537 LOC mixing four concerns (color primitives,
list/tree rendering, detail rendering, comparison rendering). Split
into four focused files of 90-190 LOC each plus a 10-line doc.go.

No functional changes.
